refactor(api): encapsulate all wiring in api.Init — internal.go no longer knows v0 or hub
